+++ src/eca/llm_providers/anthropic.clj
@@ -155,7 +155,6 @@
    {:keys [on-message-received on-error on-reason on-prepare-tool-call on-tools-called on-usage-updated]}]
   (let [messages (concat (normalize-messages past-messages supports-image?)
                          (normalize-messages (fix-non-thinking-assistant-messages user-messages) supports-image?))
-        _ (logger/info "-->" messages)
         body (merge (assoc-some
                      {:model model
                       :messages (add-cache-to-last-message messages)